+++ src/Command/AbstractCommand.php
@@ -21,13 +21,13 @@ abstract class AbstractCommand extends Command
 {
     /** @var Config */
     protected $config = null;
-    
+
     /** @var AdapterInterface */
     protected $adapter;
-    
+
     /** @var Manager */
     protected $manager;
-    
+
     /**
      * @param string $name
      * @return AbstractCommand
@@ -39,14 +39,14 @@ abstract class AbstractCommand extends Command
         }
         return $this;
     }
-    
+
     protected function configure()
     {
         $this->addOption('environment', 'e', InputOption::VALUE_REQUIRED);
         $this->addOption('config', 'c', InputOption::VALUE_OPTIONAL, 'Path to config file');
         $this->addOption('config_type', 't', InputOption::VALUE_OPTIONAL, 'Type of config, available values: php, yml, neon');
     }
-    
+
     /**
      * @param array $configuration
      * @return AbstractCommand
@@ -56,7 +56,7 @@ abstract class AbstractCommand extends Command
         $this->config = new Config($configuration);
         return $this;
     }
-    
+
     /**
      * @param InputInterface $input
      * @param OutputInterface $output
@@ -64,36 +64,61 @@ abstract class AbstractCommand extends Command
     final protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->loadConfig($input);
-        
+
         $environment = $input->getOption('environment') ?: $this->config->getDefaultEnvironment();
         $this->adapter = AdapterFactory::instance($this->config->getEnvironmentConfig($environment));
-        
+
         $this->manager = new Manager($this->config, $this->adapter);
         $this->check($input, $output);
-        
+
         $start = microtime(true);
         $this->runCommand($input, $output);
         $output->writeln('');
         $output->write('<comment>All done. Took ' . sprintf('%.4fs', microtime(true) - $start) . '</comment>');
         $output->writeln('');
     }
-    
+
     private function loadConfig(InputInterface $input)
     {
         if ($this->config) {
             return;
         }
-        $configFile = $input->getOption('config') ?: 'phoenix.php';
+
+        $configFile = $input->getOption('config');
+        if (!$configFile) {
+            $configFile = $this->getDefaultConfig();
+        }
+
+        if (!$configFile) {
+            throw new ConfigException('No configuration file exists. Create phoenix.php or phoenix.yml or phoenix.neon or phoenix.json in your project root or specify path to your existing config file with --config option');
+        }
+
         if ($configFile && !file_exists($configFile)) {
             throw new ConfigException('Configuration file "' . $configFile . '" doesn\'t exist.');
         }
-        
+
         $type = $input->getOption('config_type') ?: pathinfo($configFile, PATHINFO_EXTENSION);
         $configParser = ConfigParserFactory::instance($type);
         $configuration = $configParser->parse($configFile);
         $this->config = new Config($configuration);
     }
-    
+
+    private function getDefaultConfig()
+    {
+        $defaultConfigFiles = [
+            'phoenix.php',
+            'phoenix.yml',
+            'phoenix.neon',
+            'phoenix.json',
+        ];
+        foreach ($defaultConfigFiles as $defaultConfigFile) {
+            if (file_exists($defaultConfigFile)) {
+                return $defaultConfigFile;
+            }
+        }
+        return null;
+    }
+
     private function check(InputInterface $input, OutputInterface $output)
     {
         try {
@@ -106,11 +131,11 @@ abstract class AbstractCommand extends Command
                 $init->execute($input, $output);
             }
         }
-        
+
         if ($executedMigrations !== false && $this instanceof InitCommand) {
             throw new WrongCommandException('Phoenix was already initialized, run migrate or rollback command now.');
         }
     }
-    
+
     abstract protected function runCommand(InputInterface $input, OutputInterface $output);
 }