+++ src/Database/QueryBuilder/PgsqlQueryBuilder.php
@@ -11,13 +11,16 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
     protected $typeMap = [
         Column::TYPE_STRING => 'varchar(%d)',
         Column::TYPE_INTEGER => 'int4',
+        Column::TYPE_BIG_INTEGER => 'int8',
         Column::TYPE_BOOLEAN => 'bool',
         Column::TYPE_TEXT => 'text',
+        Column::TYPE_DATE => 'date',
         Column::TYPE_DATETIME => 'timestamp(6)',
         Column::TYPE_UUID => 'uuid',
         Column::TYPE_JSON => 'json',
         Column::TYPE_CHAR => 'char(%d)',
         Column::TYPE_DECIMAL => 'decimal(%d,%d)',
+        Column::TYPE_FLOAT => 'real',
     ];
     
     protected $defaultLength = [